[3.0b2] -=Backup worldConfig=- untested version!
- Revamped ZoneVar, also use it for WorldConfig now.
- Extended ZoneVar to allow for all configuration options of WorldConfig
- Completely removed accessresolvers, powerfull system but way too complicated which made adding features harder.
  I also never needed the power what the system offered.
- Cleaned up some code because of the accessresolver removal.
- First work on zimport, not finished yet.
- Removed Textiel, not used anymore.
- Added new WorldManager based on Prtree.
- Switched y and z around(synced to minecraft now).
- Protection Stone like support.
- added tests.
- bugfixes.

+++ src/com/zones/model/ZonesAccess.java
@@ -78,7 +78,7 @@ public class ZonesAccess {
             if(a == null) {
                 continue;
             }
-            r = r | a;
+            r = r | a.intValue();
         }
         return all[r];
     }
